Removed the "PW" prefix from all the Application level files.  (I've kept some of them on common files which I intend to be part of a library at some future date. But the model and view-controller classes have been renamed).

+++ Stereogram/PWAlertView.m
@@ -70,7 +70,7 @@
     
     for (UIAlertAction *action in self.actions) {
             // Note: Destructive events are not shown specially here. Alerts only show for destructive events anyway so all actions are assumed to be destructive or cancel.
-        int index = [_alertView addButtonWithTitle:action.title];
+        NSInteger index = [_alertView addButtonWithTitle:action.title];
         if (action.style == UIAlertActionStyleCancel) {
             _alertView.cancelButtonIndex = index;
         }
